requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
abseil-cpp=20211102.0=h27087fc_1
aiohttp=3.8.1=py37h540881e_1
aiosignal=1.2.0=pyhd8ed1ab_0
arrow-cpp=0.11.1=py37h3bd774a_1
async-timeout=4.0.2=pyhd8ed1ab_0
asynctest=0.13.0=py_0
attrs=21.4.0=pyhd8ed1ab_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=py_2
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
boost-cpp=1.68.0=h11c811c_1000
bottleneck=1.3.5=py37hda87dfa_0
brotlipy=0.7.0=py37h540881e_1004
bzip2=1.0.8=h7f98852_4
c-ares=1.18.1=h7f98852_0
ca-certificates=2022.6.15=ha878542_0
cachetools=4.2.4=pyhd8ed1ab_0
certifi=2022.6.15=py37h89c1867_0
cffi=1.15.1=py37h43b0acd_0
charset-normalizer=2.1.0=pyhd8ed1ab_0
click=8.1.3=pypi_0
colorama=0.4.5=pyhd8ed1ab_0
cryptography=37.0.2=py37h38fbfac_0
decorator=5.1.1=pyhd8ed1ab_0
emoji=2.1.0=pypi_0
filelock=3.8.0=pypi_0
frozenlist=1.3.0=py37h540881e_1
google-api-core=1.22.2=py37hc8dfbb8_0
google-auth=1.35.0=pyh6c4a22f_0
google-cloud-bigquery=1.28.0=pyhc8dfbb8_0
google-cloud-bigquery-core=1.28.0=py37hc8dfbb8_0
google-cloud-bigquery-storage=0.6.0=py37_0
google-cloud-core=1.7.1=pyhd3eb1b0_0
google-crc32c=1.1.2=py37h5d4fa31_3
google-resumable-media=1.3.1=pyhd3eb1b0_1
googleapis-common-protos=1.56.3=py37h89c1867_0
greenlet=1.1.3=pypi_0
grpc-cpp=1.47.0=hbd84cd8_1
grpcio=1.47.0=py37h64c7598_1
huggingface-hub=0.10.0=pypi_0
icu=58.2=hf484d3e_1000
idna=3.3=pyhd8ed1ab_0
importlib-metadata=5.0.0=pypi_0
ipython=7.33.0=py37h89c1867_0
jedi=0.18.1=py37h89c1867_1
joblib=1.2.0=pypi_0
ld_impl_linux-64=2.36.1=hea4e1c9_2
libblas=3.9.0=15_linux64_openblas
libcblas=3.9.0=15_linux64_openblas
libcrc32c=1.1.2=h9c3ff4c_0
libevent=2.1.10=h9b69904_4
libffi=3.4.2=h7f98852_5
libgcc-ng=12.1.0=h8d9b700_16
libgfortran-ng=12.1.0=h69a702a_16
libgfortran5=12.1.0=hdcd56e2_16
libgomp=12.1.0=h8d9b700_16
liblapack=3.9.0=15_linux64_openblas
libnsl=2.0.0=h7f98852_0
libopenblas=0.3.20=pthreads_h78a6416_0
libprotobuf=3.20.1=h6239696_0
libstdcxx-ng=12.1.0=ha89aaad_16
libzlib=1.2.12=h166bdaf_1
matplotlib-inline=0.1.3=pyhd8ed1ab_0
multidict=6.0.2=py37h540881e_1
ncurses=6.3=h27087fc_1
nltk=3.7=pypi_0
nomkl=1.0=h5ca1d4c_0
numexpr=2.8.0=py37h85a3170_102
numpy=1.19.5=py37h3e96413_3
openssl=1.1.1q=h166bdaf_0
packaging=21.3=pypi_0
pandas=1.3.5=py37h8c16a72_0
parquet-cpp=1.5.1=3
parso=0.8.3=pyhd8ed1ab_0
pexpect=4.8.0=pyh9f0ad1d_2
pickleshare=0.7.5=py_1003
pillow=9.2.0=pypi_0
pip=22.1.2=pyhd8ed1ab_0
prompt-toolkit=3.0.30=pyha770c72_0
protobuf=3.20.1=py37hd23a5d3_0
ptyprocess=0.7.0=pyhd3deb0d_0
pyarrow=0.11.1=py37hbbcf98d_1002
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.7=py_0
pycparser=2.21=pyhd8ed1ab_0
pygments=2.12.0=pyhd8ed1ab_0
pyopenssl=22.0.0=pyhd8ed1ab_0
pyparsing=3.0.9=pypi_0
pysocks=1.7.1=py37h89c1867_5
python=3.7.12=hb7a2778_100_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.7=2_cp37m
pytz=2022.1=pyhd8ed1ab_0
pyu2f=0.1.5=pyhd8ed1ab_0
pyyaml=6.0=pypi_0
re2=2022.06.01=h27087fc_0
readline=8.1.2=h0f457ee_0
regex=2022.9.13=pypi_0
requests=2.28.1=pyhd8ed1ab_0
rsa=4.8=pyhd8ed1ab_0
scikit-learn=1.0.2=pypi_0
scipy=1.7.3=pypi_0
sentence-transformers=2.2.2=pypi_0
sentencepiece=0.1.97=pypi_0
setuptools=63.1.0=py37h89c1867_0
six=1.16.0=pyh6c4a22f_0
sqlalchemy=1.4.41=pypi_0
sqlite=3.39.0=h4ff8645_0
stanza=1.4.2=pypi_0
threadpoolctl=3.1.0=pypi_0
thrift-cpp=0.11.0=h02b749d_3
tk=8.6.12=h27826a3_0
tokenizers=0.12.1=pypi_0
torch=1.12.1=pypi_0
torchvision=0.13.1=pypi_0
tqdm=4.64.0=pyhd8ed1ab_0
traitlets=5.3.0=pyhd8ed1ab_0
transformers=4.22.2=pypi_0
typing-extensions=4.3.0=hd8ed1ab_0
typing_extensions=4.3.0=pyha770c72_0
urllib3=1.26.9=pyhd8ed1ab_0
wcwidth=0.2.5=pyh9f0ad1d_2
wheel=0.37.1=pyhd8ed1ab_0
xz=5.2.5=h516909a_1
yarl=1.7.2=py37h540881e_2
zipp=3.8.1=pypi_0
zlib=1.2.12=h166bdaf_1
